+++ appseed/ca/multithreading/multithreading_wait_result.h
@@ -38,7 +38,7 @@ public:
 	///  \brief		constructor with result value and object count passed (Windows only)
 	///  \param		value waiting result value as passed by Windows API function
 	///  \param		objectCount number of items in event_collection
-	explicit wait_result(uint32_t long value, count objectCount=1 );
+	explicit wait_result(uint32_t value, count objectCount = 1);
 //#endif
 	///  \brief		constructor with result value passed
 	///  \param		value waiting result value as passed by Windows API function
